+++ src/main/java/blackjack/view/OutputView.java
@@ -1,12 +1,12 @@
 package blackjack.view;
 
-import blackjack.dto.Participants;
-import blackjack.dto.GameResult;
 import blackjack.domain.ResultType;
 import blackjack.domain.card.Card;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Player;
+import blackjack.dto.GameResult;
+import blackjack.dto.Participants;
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -18,7 +18,8 @@ public class OutputView {
         System.out.println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)");
     }
 
-    public static void printGameInitializeMessage(Participants participants, int startingCardCount) {
+    public static void printGameInitializeMessage(Participants participants,
+        int startingCardCount) {
         String participantNames = participants.unwrap().stream()
             .map(Participant::getName)
             .collect(Collectors.joining(NAME_DELIMITER));
@@ -26,7 +27,8 @@ public class OutputView {
     }
 
     public static void printParticipantsStatus(Participants participants, boolean withScore) {
-        participants.unwrap().forEach(participant -> printParticipantStatus(participant, withScore));
+        participants.unwrap()
+            .forEach(participant -> printParticipantStatus(participant, withScore));
         System.out.println();
     }
