+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,14 +1,16 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.Deck;
+import blackjack.domain.Hand;
+import blackjack.domain.Response;
+import blackjack.domain.ResultType;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardValue;
 import blackjack.domain.card.Shape;
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Player;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
